Rename message layout text to inline
